+++ src/game/server/ddnetpp/lua/table_unpacker.cpp
@@ -228,31 +228,31 @@ vec2 CTableUnpacker::GetPosition(const char *pKey)
 	return LuaPosToServer(GetVec2(pKey));
 }
 
-int CTableUnpacker::GetCoordinate(const char *pKey)
+float CTableUnpacker::GetCoordinate(const char *pKey)
 {
 	if(m_IsError)
-		return 0;
+		return 0.0f;
 	lua_getfield(LuaState(), m_Index, pKey);
 	if(lua_isnoneornil(LuaState(), -1))
 	{
 		char aBuf[512];
 		str_format(aBuf, sizeof(aBuf), "missing key '%s'", pKey);
 		Error(aBuf);
-		return 0;
+		return 0.0f;
 	}
 	if(!lua_isnumber(LuaState(), -1) && !lua_isinteger(LuaState(), -1))
 	{
 		char aBuf[512];
 		str_format(aBuf, sizeof(aBuf), "expected key '%s' to hold a number", pKey);
 		Error(aBuf);
-		return 0;
+		return 0.0f;
 	}
-	int Val = LuaCheckCoordinate(LuaState(), -1);
+	float Val = LuaCheckCoordinate(LuaState(), -1);
 	lua_pop(LuaState(), 1);
 	return Val;
 }
 
-std::optional<int> CTableUnpacker::GetCoordinateOptional(const char *pKey)
+std::optional<float> CTableUnpacker::GetCoordinateOptional(const char *pKey)
 {
 	if(m_IsError)
 		return std::nullopt;
@@ -267,7 +267,7 @@ std::optional<int> CTableUnpacker::GetCoordinateOptional(const char *pKey)
 		lua_pop(LuaState(), 1);
 		return std::nullopt;
 	}
-	int Val = LuaCheckCoordinate(LuaState(), -1);
+	float Val = LuaCheckCoordinate(LuaState(), -1);
 	lua_pop(LuaState(), 1);
 	return Val;
 }
